remove global state and add inline loaders to get the state
