+++ ra-trustvox/components/TrustvoxCertificateFixed.tsx
@@ -1,0 +1,5 @@
+export default function TrustvoxCertificateFixed() {
+  return (
+    <div data-trustvox-certificate-fixed="data-trustvox-certificate-fixed" />
+  );
+}
