+++ ra-trustvox/manifest.gen.ts
@@ -2,15 +2,17 @@
 // This file SHOULD be checked into source version control.
 // This file is automatically updated during development when running `dev.ts`.
 
-import * as $$$$$$0 from "./sections/TrustvoxProductReviews.tsx";
-import * as $$$$$$1 from "./sections/TrustvoxRateConfig.tsx";
-import * as $$$$$$2 from "./sections/TrustvoxStoreReviewsCarousel.tsx";
+import * as $$$$$$0 from "./sections/TrustvoxCertificate.tsx";
+import * as $$$$$$1 from "./sections/TrustvoxProductReviews.tsx";
+import * as $$$$$$2 from "./sections/TrustvoxRateConfig.tsx";
+import * as $$$$$$3 from "./sections/TrustvoxStoreReviewsCarousel.tsx";
 
 const manifest = {
   "sections": {
-    "ra-trustvox/sections/TrustvoxProductReviews.tsx": $$$$$$0,
-    "ra-trustvox/sections/TrustvoxRateConfig.tsx": $$$$$$1,
-    "ra-trustvox/sections/TrustvoxStoreReviewsCarousel.tsx": $$$$$$2,
+    "ra-trustvox/sections/TrustvoxCertificate.tsx": $$$$$$0,
+    "ra-trustvox/sections/TrustvoxProductReviews.tsx": $$$$$$1,
+    "ra-trustvox/sections/TrustvoxRateConfig.tsx": $$$$$$2,
+    "ra-trustvox/sections/TrustvoxStoreReviewsCarousel.tsx": $$$$$$3,
   },
   "name": "ra-trustvox",
   "baseUrl": import.meta.url,
